tests/submission/Infraware/Forms/contents/Forms/Not Use/input_max.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>input_max</h3>
	</p>
	
	<li>Max value of number is 30, You cannot submit this form.</li>
	
	<hr>

	<div id="log"></div>
	
	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  name="input_form">
	<p><input type='number' id='input_number'></p>
	</form>

	<script>
		
		var input_number = document.getElementById("input_number");
		input_number.max = "30";

		if (typeof(input_number.max) == "string") {
			if (input_number.max != "30") {
				test(function() {
					assert_true(false, "max attribute is not correct.");
				});
			}	
		} else {
			test(function() {
				assert_unreached("max attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
